Move login logic to LoginService
